+++ sdk/cosmos/azure-cosmos/azure/cosmos/_service_response_retry_policy.py
@@ -10,7 +10,6 @@ only do cross regional retries for read operations.
 
 import logging
 from azure.cosmos.documents import _OperationType
-from azure.cosmos._base import try_ppaf_failover_threshold
 
 class ServiceResponseRetryPolicy(object):
 
@@ -51,7 +50,7 @@ class ServiceResponseRetryPolicy(object):
         if self.request:
             # We track consecutive failures for per partition automatic failover, and only fail over at a partition
             # level after the threshold is reached
-            try_ppaf_failover_threshold(self.global_endpoint_manager, self.pk_range_wrapper, self.request)
+            self.global_endpoint_manager.try_ppaf_failover_threshold(self.pk_range_wrapper, self.request)
             if not _OperationType.IsReadOnlyOperation(self.request.operation_type) and not self.request.retry_write:
                 return False
             if self.request.retry_write and self.failover_retry_count + 1 >= self.max_write_retry_count: