+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_retry_utility_async.py
@@ -202,11 +202,9 @@ async def ExecuteAsync(client, global_endpoint_manager, function, *args, **kwarg
                     if retry_policy.should_update_throughput_link(request.body, cached_container):
                         new_body = retry_policy._update_throughput_link(request.body)
                         request.body = new_body
-
                     retry_policy.container_rid = cached_container["_rid"]
                     request.headers[retry_policy._intended_headers] = retry_policy.container_rid
             elif e.status_code == StatusCodes.SERVICE_UNAVAILABLE:
-                # if ppaf is applicable, we record the failure
                 retry_policy = service_unavailable_retry_policy
             elif e.status_code == StatusCodes.REQUEST_TIMEOUT or e.status_code >= StatusCodes.INTERNAL_SERVER_ERROR:
                 if args: